+++ .github/workflows/python.yml
@@ -18,6 +18,6 @@ jobs:
         run: pip install .
       - name: Try to run the command
         run: |
-          fedcloud-vo-monitor --help
+          fedcloud-vm-monitor --help
           fedcloud-sla-monitor --help
           fedcloud-vo-testing --help
